+++ src/lib.rs
@@ -32,7 +32,7 @@
 use crate::{
     addresses::PublicAddresses,
     config::Litep2pConfig,
-    error::DialError,
+    error::{DialError, DnsInitError},
     protocol::{
         libp2p::{
             bitswap::Bitswap,
@@ -59,7 +59,11 @@ use crate::transport::webrtc::WebRtcTransport;
 #[cfg(feature = "websocket")]
 use crate::transport::websocket::WebSocketTransport;
 
-use hickory_resolver::{name_server::TokioConnectionProvider, TokioResolver};
+use hickory_resolver::{
+    config::{LookupIpStrategy, ResolverConfig, ResolverOpts, GOOGLE},
+    net::runtime::TokioRuntimeProvider,
+    TokioResolver,
+};
 use multiaddr::{Multiaddr, Protocol};
 use transport::Endpoint;
 use types::ConnectionId;
@@ -163,16 +167,22 @@ impl Litep2p {
         let bandwidth_sink = BandwidthSink::new();
         let mut listen_addresses = vec![];
 
-        let (resolver_config, resolver_opts) = if litep2p_config.use_system_dns_config {
+        let (resolver_config, mut resolver_opts) = if litep2p_config.use_system_dns_config {
             hickory_resolver::system_conf::read_system_conf()
-                .map_err(Error::CannotReadSystemDnsConfig)?
+                .map_err(|error| Error::CannotReadSystemDnsConfig(DnsInitError::new(error)))?
         } else {
-            (Default::default(), Default::default())
+            (
+                ResolverConfig::udp_and_tcp(&GOOGLE),
+                ResolverOpts::default(),
+            )
         };
+        resolver_opts.ip_strategy = LookupIpStrategy::Ipv4AndIpv6; // prefer A records
+
         let resolver = Arc::new(
-            TokioResolver::builder_with_config(resolver_config, TokioConnectionProvider::default())
+            TokioResolver::builder_with_config(resolver_config, TokioRuntimeProvider::default())
                 .with_options(resolver_opts)
-                .build(),
+                .build()
+                .map_err(|error| Error::DnsResolverInit(DnsInitError::new(error)))?,
         );
 
         let supported_transports = Self::supported_transports(&litep2p_config);